+++ lib/src/api/account_api.dart
@@ -43,6 +43,12 @@ class AccountApi {
         (json) => StickerAlbum.fromJson(json),
       );
 
+  Future<MixinResponse<StickerAlbum>> getStickerAlbum(String id) =>
+      MixinResponse.request(
+        dio.get('/albums/$id'),
+        (json) => StickerAlbum.fromJson(json),
+      );
+
   Future<MixinResponse<List<Sticker>>> getStickersByAlbumId(String id) =>
       MixinResponse.requestList(
         dio.get('/stickers/albums/$id'),
@@ -84,5 +90,18 @@ class AccountApi {
           }
           return null;
         },
+      );
+
+  Future<MixinResponse<VerificationResponse>> verification(
+          VerificationRequest request) =>
+      MixinResponse.request<VerificationResponse>(
+        dio.post('/verifications', data: request),
+        (json) => VerificationResponse.fromJson(json),
+      );
+
+  Future<MixinResponse<Account>> create(String id, AccountRequest request) =>
+      MixinResponse.request<Account>(
+        dio.post('/verifications/$id', data: request),
+        (json) => Account.fromJson(json),
       );
 }