+++ lib/src/vo/message/live_message.dart
@@ -4,6 +4,7 @@ part 'live_message.g.dart';
 
 @JsonSerializable()
 class LiveMessage {
+  // ignore: avoid_positional_boolean_parameters
   LiveMessage(this.width, this.height, this.thumbUrl, this.url, this.shareable);
 
   factory LiveMessage.fromJson(Map<String, dynamic> json) =>