+++ lib/src/vo/payment_response.g.dart
@@ -8,39 +8,13 @@ part of 'payment_response.dart';
 
 PaymentResponse _$PaymentResponseFromJson(Map<String, dynamic> json) =>
     PaymentResponse(
-      status: _$enumDecode(_$PaymentStatusEnumMap, json['status']),
+      status: $enumDecode(_$PaymentStatusEnumMap, json['status']),
     );
 
 Map<String, dynamic> _$PaymentResponseToJson(PaymentResponse instance) =>
     <String, dynamic>{
       'status': _$PaymentStatusEnumMap[instance.status],
     };
-
-K _$enumDecode<K, V>(
-  Map<K, V> enumValues,
-  Object? source, {
-  K? unknownValue,
-}) {
-  if (source == null) {
-    throw ArgumentError(
-      'A value must be provided. Supported values: '
-      '${enumValues.values.join(', ')}',
-    );
-  }
-
-  return enumValues.entries.singleWhere(
-    (e) => e.value == source,
-    orElse: () {
-      if (unknownValue == null) {
-        throw ArgumentError(
-          '`$source` is not one of the supported values: '
-          '${enumValues.values.join(', ')}',
-        );
-      }
-      return MapEntry(unknownValue, enumValues.values.first);
-    },
-  ).key;
-}
 
 const _$PaymentStatusEnumMap = {
   PaymentStatus.pending: 'pending',